页面设计；增加了置信度-提升度图表；后台返回支持度改为支持度计数

+++ src/main/java/com/lbh/dpdfim/ui/service/FrequentItemSetDTO.java
@@ -7,6 +7,6 @@ public class FrequentItemSetDTO {
 
     private String items;
 
-    private double support;
+    private Long support;
 
 }
